Add a test for get_closest_point
Check that the returned offset actually matches the point
that is returned by gsk_path_measure_get_closest_point_full.

We do the extra work to detect when we end up with a point
that is straddling a contour boundary, and accept if get_point
returns a different point than get_closest_point in this case.